GameEngine nightly - The first big update

+++ src/demos/town/buildings/houses/House.java
@@ -14,6 +14,8 @@ import game.IO.load.LoadRequest;
  *
  */
 public class House extends Building {
+	
+	//FIXME: Remove this and move this into Building
 
 	protected BufferedImage outline, building, built;
 
@@ -23,7 +25,7 @@ public class House extends Building {
 	 * @param width
 	 * @param height
 	 */
-	public House(float x, float y, float width, float height) {
+	public House(float x, float y, int width, int height) {
 		super(x, y, width, height);
 		load();
 	}
